libsepol/fuzz: be more strict
Validate the policy after optimization and validate writing back in
binary format is successful.  This treats OOM failures as critical,
which hopefully is not an issue with the OSS-FUZZ infrastructure.
Writing policies in conf or CIL format might still fail, e.g. due to
unsupported policy capabilities.

Set the O_CLOEXEC flag for /dev/null, in the case of the existence of
sibling threads.

Signed-off-by: Christian Göttsche <cgzones@googlemail.com>
---
TODO:
   should policy validation reject unsupported (newer) policy
   capabilities?